+++ NEWS.md
@@ -6,6 +6,7 @@ Version 0.11a0 - dev
 --------------------
 
 - NEW: `Dimension.get_measurement()` supports angular, angular3p and ordinate dimensions
+- NEW: `Drawing.set_modelspace_vport()` set initial view/zoom location for the modelspace
 
 Version 0.10.2 - 2019-10-05
 ---------------------------